refactor(model): give each inverse-design search its own module
The two searches share the target vocabulary and nothing else — 341 and 489 lines sitting in one
class because that is where they were written. They move to latent.py and composition.py as
functions taking the model as their first argument, which states the dependency instead of
assuming it and is what let them leave the class at all.

They are bound back as class attributes rather than wrapped. A plain function assigned in a class
body binds as a method, so model.optimize_composition(...) is unchanged for every caller and every
keyword keeps its type. A def forwarding *args/**kwargs would have been shorter and would have
erased both — the twenty-keyword signature and mypy's ability to check a call against it.

Verified the same way as the earlier moves: an AST-level comparison against the methods they
replace. optimize_composition is token-identical over 439 non-blank lines; optimize_latent differs
only where ruff rejoined one raise onto a single line once it was dedented far enough to fit.

    mixin.py   1118 -> 291

What the class keeps is what actually needs a model: the four-member contract, and the two steps
that turn a target into a loss term against real heads.
